[Maps] Move vector-style component to vector folder (#51779)
