+++ x-pack/legacy/plugins/maps/public/layers/styles/_index.scss
@@ -1,3 +1,3 @@
 @import './components/color_gradient';
-@import './components/static_dynamic_style_row';
+@import './vector/components/static_dynamic_style_row';
 @import './vector/components/color/color_stops';
