+++ x-pack/legacy/plugins/maps/public/layers/styles/vector/components/_static_dynamic_style_row.scss
