+++ x-pack/legacy/plugins/maps/public/layers/styles/vector/components/size/vector_style_size_editor.js
@@ -6,7 +6,7 @@
 
 import React from 'react';
 
-import { StaticDynamicStyleRow } from '../../../components/static_dynamic_style_row';
+import { StaticDynamicStyleRow } from '../static_dynamic_style_row';
 import  { DynamicSizeSelection } from './dynamic_size_selection';
 import  { StaticSizeSelection } from './static_size_selection';
 import { getVectorStyleLabel } from '../get_vector_style_label';
